% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_earth.R
\name{step_earth}
\alias{step_earth}
\title{Step earth}
\usage{
step_earth(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  drop = TRUE,
  outcome = NULL,
  options = list(),
  res = NULL,
  prefix = "EARTH",
  skip = FALSE,
  id = recipes::rand_id("earth")
)
}
\arguments{
\item{recipe}{recipe}

\item{...}{selector}

\item{role}{role}

\item{trained}{trained}

\item{drop}{drop}

\item{outcome}{outcome}

\item{options}{options}

\item{res}{res}

\item{prefix}{prefix}

\item{skip}{skip}

\item{id}{id}
}
\description{
Step earth
}
